# More efficient decoding of floats and doubles with long mantissas