… zio-json/native/src/main/scala/zio/json/internal/UnsafeNumbers.scala