Tested with Scala 3 and JDK 21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfBigDecimalsReading.zioJson     128  thrpt    5   139494.914 ±  4433.385  ops/s
ArrayOfDoublesReading.zioJson         128  thrpt    5   239723.251 ± 22115.957  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   369198.260 ± 18019.799  ops/s
```

#### After
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfBigDecimalsReading.zioJson     128  thrpt    5   151187.739 ±   459.966  ops/s
ArrayOfDoublesReading.zioJson         128  thrpt    5   304200.753 ±  4779.653  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   443325.617 ±  6852.218  ops/s
```